add a factor that should decrease the error caused by missing peaks changed latest bpm result to be CGFloat

+++ Heartbeat/Algorithm.h
@@ -25,7 +25,7 @@
 
 @property (nonatomic , readonly) BOOL isCalibrationOver;
 @property (nonatomic , readonly) BOOL isFinalResultDetermined;
-@property (nonatomic , readonly) NSUInteger bpmLatestResult;
+@property (nonatomic , readonly) CGFloat bpmLatestResult;
 
 //
